requirements.txt
# This file was autogenerated by uv via the following command:
#    uv pip compile pyproject.toml -o requirements.txt
aiohappyeyeballs==2.6.1
    # via aiohttp
aiohttp==3.13.4
    # via livekit-api
aiosignal==1.4.0
    # via aiohttp
attrs==26.1.0
    # via aiohttp
frozenlist==1.8.0
    # via
    #   aiohttp
    #   aiosignal
idna==3.11
    # via yarl
livekit-api==1.1.0
    # via livekit-local (pyproject.toml)
livekit-protocol==1.1.3
    # via livekit-api
multidict==6.7.1
    # via
    #   aiohttp
    #   yarl
propcache==0.4.1
    # via
    #   aiohttp
    #   yarl
protobuf==7.34.1
    # via
    #   livekit-api
    #   livekit-protocol
pyjwt==2.12.1
    # via livekit-api
types-protobuf==6.32.1.20260221
    # via
    #   livekit-api
    #   livekit-protocol
yarl==1.23.0
    # via aiohttp
